client: move fetchWithProgress into xhr-util importable file
